implement basic timer support
This commit updates for recent changes that eliminate the need to
use xargo for arm and implements basic support for TC3.

I've measured the logic levels with my nscope and this appears to
be accurately respecting the time configured, but I can't really
measure this beyond a few KHz.

It's certainly good enough to blink the red led!