bump to 0.2 and prep to take over atsamd21g18a on crates.io
I didn't realize that there was already a package.  I've
been communicating with its owner and we're good to continue
with my fork.

The version needs to bump up to 0.2 to reflect the interface
changes that have happened as a result of changes in svd2rust
since the original package was uploaded.