GenericClockController::new -> with_internal_32kosc
The internal 32kosc is not sufficient to use for USB so we need
to make sure that we have a way to initialize the clocks
with an external oscillator if one is present.

Before adding support for that, rename the current constructor
to reflect what how the system is currently being initialized.