Make doc build more efficient
Restructure the doc build so that it only builds crates from this
repo, and uses workspaces to make the doc build itself more efficient.

Note that this removes the default PAC selection from the HAL crate.
You may need to `cargo clean` to correctly invalidate your build.

This also means that `cargo package` for the hal crate will need
to be run using `--no-verify` to skip trying to build the crate
with no PAC feature enabled.