trellis_m4: use ws2812-nop-samd51, disable lto
We measured ws2812-nop-samd51's bitbanged signal to the NeoPixels  with
a logic analyzer and discovered it was going too fast with LTO enabled
(which was needed originally to get the timer-driven examples working).

With LTO off ws2812-nop-samd51 can *correctly* drive the NeoPixels,
and the examples no longer run much too fast like they did with the
timer-driven approach.