hal: dma::controller::Registers trait
Defines a `Registers` trait which abstracts over the DMAC registers,
allowing DMA drivers to be written targeting this trait, and then used
with any relevant PAC which already maps them with svd2rust.

The trait is defined differently for SAMD21 vs SAMD51 as they have
different DMAC registers, but usable with both. An additional
higher-level abstraction trait will have to be written to abstract
across the differences between the two.

This approach is relatively macro-heavy, but the result should enable
the rest of the DMA subsystem to be written sans macros by leveraging
the newly defined trait.