Refactor DMA controller into SAMD21/SAMD51 types
This replaces all of the SAMD21/SAMD51 conditional compilation with
separate `dma::controller::ControllerCritical` structs for each of the
respective MCUs that handle the differences between them.

It also wraps DMAC in a `cortex_m::interrupt::Mutex` and uses that to
handle all locking associated with accessing the DMAC registers and
channel mask.

Additionally this deletes the "not recommended" suspend/resume
functionality, as the Adafruit comments already suggest it was
misunderstood and should not be used.