Adds UART setup for all Adafruit boards
Pin parameters to uart functions are named after the names in
Pins struct so that it’s clearer to find out what needs to be
passed in.

Not directly tested because I don’t have access to all of the
boards. Pin info was taken from variant.cpp files and SERCOM
type safety ensures that the rest is set up correctly.

Also tweaks Trellis M4 to be "STEMMA" instead of "JST" since that’s
what it’s called in pygamer.