Use bleeding-edge `svd2rust`
A pair of changes (rust-embedded/svd#119, rust-embedded/svd2rust#444)
recently went into the `svd2rust` and `svd-parser` crates that improves
the handling of cluster names and the generation of full type paths for
cluster instances. Until these changes are part of a release have
`update.sh` pull a fork of `svd2rust` that uses the master version of
`svd-parser`.

See #195.