Allow timers to use period instead of frequency (#207)
* Convert timers to using period vs frequency

By using frequency to track a timer's period, you limit yourself in the possible timer options. For one, a timer longer than 1s is not possible, and many fractional periods are not possible. There are some instances where frequency works a little better, for example 3 Hz which has a rounding error if you convert to microseconds, but that error is extremely small.

I was able to test this on a feather_m0 board, but not the other boards.

* src_freq is u32, not Hertz

* missed some changes to panic

* new_hz -> new to maintain API