fix(pwm): Limit pwm re-export to unproven
The SAMD51 and SAME54 HALs re-export the `pwm` module to avoid breaking
the previous, public API. The `pwm` module is currently unproven, but
the re-export occurred regardless, creating an error when building
without the `unproven` feature. Add a `cfg` attribute to limit the
re-export to builds with the `unproven` feature.