Refactor the sercom module and pads submodule
Combine the SAMD11 and SAMD21 sercom modules. Restructure the sercom
modules to use v1 and v2 submodules like the gpio module. Add a new
sercom::v2::pads module that doesn't rely as heavily on macros. Replace
the sercom::v1::pads module with a shim the uses the v2::pads module
under the hood.