Add Pin methods to AnyPin trait and update documentation
Although there is a one-to-one mapping between Pin<I, M> and
AnyPin<Id = I, Mode = M>, the compiler doesn't know that. Provide a
means to convert between the two, and implement all of Pin's public
methods in AnyPin.

Update the documentation for AnyPin and update intra-doc links to use
monospaced font.