Simplify GPIO pin mode trait bounds
Currently, the PinMode trait is implemented on each individual mode,
e.g. Input<Floating> or Output<PushPull>. By doing so, the compiler is
unable to infer that Output<C>, for example, implements PinMode for all
C, because there is no declaration to that effect. As a consequence, the
compiler often requires seemingly superfluous trait bounds like
`Output<C>: PinMode`.

Restructure the Config and PinMode traits to allow implementations of
the form `impl<C: OutputConfig> PinMode for Output<C>`. This eliminates
the need for the extra trait bound.