Add conversion between v1 and v2 SERCOM pads
Add implementations of From between v1 and v2 SERCOM pads for thumbv6m
targets. Doing the same for thumbv7em targets is more difficult. The
implementation overlaps with that of the thumbv6m conversion, and IOSET
information is lost when going from a v2 Pad to a v1 Pad. For these
reasons, this conversion is excluded. As a workaround, it is always
possible to use pad.free() to recover the pin and do the conversion
manually.