Improve gpio::v2 in a few areas (#349)
* Add additional BSP peripheral helper functions (#344)

* Add `spi_master` helper function for Trinket M0

* Add deprecation notice for usb_bus function and suggest using usb_allocator

* Add `spi_master` and `uart` helpers for wio_lite_mg126

* Hopefully make the format check on CI happy...

* Revert import changes for Metro M0 to fix build errors

* Fix a typo

* Simplify GPIO pin mode trait bounds

Currently, the PinMode trait is implemented on each individual mode,
e.g. Input<Floating> or Output<PushPull>. By doing so, the compiler is
unable to infer that Output<C>, for example, implements PinMode for all
C, because there is no declaration to that effect. As a consequence, the
compiler often requires seemingly superfluous trait bounds like
`Output<C>: PinMode`.

Restructure the Config and PinMode traits to allow implementations of
the form `impl<C: OutputConfig> PinMode for Output<C>`. This eliminates
the need for the extra trait bound.

* Improve unsafe documentation for gpio::v2

Improve the documentation for the unsafe aspects of gpio::v2.
Restructure the unsafe blocks to make proper usage more explicit.

* Add missing repr(C) to struct in gpio::v2

The GROUP struct was created to paper over difference between the PAC
structures among the SAMD11, SAMD21 and SAMx5x chips. However, it was
missing a **critical** repr(C) attribute.

Co-authored-by: Jesse Braham <jessebraham@users.noreply.github.com>