Fix gpio::v2 PORT access functions and add a bsp_pins macro (#371)
* Fix gpio::v2 PORT access functions

Previously, the gpio::v2::Pins struct provided two methods to access the
PAC PORT. One would hand out a PORT reference, while the other would
consume the Pins struct and return an owned PORT. After further
consideration, the latter function is not useful. Once you take
individual pins out of the Pins struct, you can no longer move the Pins
struct into the function to retrieve the PORT.

Instead, make the port field within the Pins struct an Option. Then,
provide a method to take back the PORT from the Pins struct after
creating it. Calling this function must be done before taking any
individual pins out of the Pins struct. The PORT can only be taken once,
and it is unsafe to do so. Subsequent attempts to take it will panic.

* Create a gpio::v2::bsp_pins macro

Create a macro to give meaningful names to pins in BSPs.

* Simplify use of bsp_pins macro

Use the HAL re-export of paste, so that users don't need to import it
themselves.

Improve the documentation to clarify importing the bsp_pins macro.