Improve the implementation of AnyPin and AnyPad (#385)
The AnyPin and AnyPad traits act as meta-types for the respective
Pin<I, M> and Pad<S, P, M> types, and they encapsulate the Pin and Pad
types with fewer type parameters.

Although useful, it was difficult to convert between a generic type
`P: AnyPin` and the corresponding `ConcretePin<P>`. Doing so required
using custom `as_concrete` and related functions. To ease this burden,
we repeated the entire `Pin` API on the `AnyPin` trait. However, that
places an additional burden on the maintainer, because any changes to
the API or documentation must be synchronized. Instead, it would be
preferrable to simplify conversion between the abstract and concrete
types, and then require users to perform that conversion manually.

Create an `Is` trait that represents type identity. The super traits of
`Is` enforce reflexive implementations of `From`, `AsRef` and `AsMut`
for `Is::Type`, meaning `T: From<T> + AsRef<T> + AsMut<T>` is always
true when `T: Is<Type = T>`. Implementations of `Is` must always have
`Is::Type == Self`.

Use the new `Is` trait to simplify converting between abstract and
generic types. Change the names `ConcretePin` and `ConcretePad` to
`SpecificPin` and `SpecificPad` to better match the names `AnyPin` and
`AnyPad`. Apply the trait bound `AnyPin: Is<Type = SpecificPin<Self>>`
and similarly for `AnyPad`. Doing so allows trivial conversion between
`P: AnyPin` and `SpecificPin<P>` using `Into`, `AsRef` and `AsMut`.

Update the documentation of the `pin` and `pads` modules to explain how
to use `Into`, `AsRef` and `AsMut` for the conversion. Also make minor
documentation updates in other areas as well.

Co-authored-by: Twitch <twitchyliquid64@users.noreply.github.com>