Provide a better abstraction for GPIO register access
Making each `Pin` and `DynPin` both `Send` and `Sync` requires some
careful considerations. Previously, this was not well abstracted and
required `unsafe` blocks spread across the modules.

Instead, consolidate the register access functions into a trait and a
pair of marker types specifically for the purpose. The new API clarifies
and codifies the safety requirements.