Update HAL to accept both v1::Pin and v2::Pins everywhere
Update all users of the `gpio` module to accept both `v1::Pin` and
`v2::Pin` types. In most cases, this was fairly straightforward and did
not involve any breaking changes. However, the `thumbv7em::pwm` required
a pretty major syntax hack. What used to be an `enum` is now a `struct`.
Update the few instances of broken code in the BSPs.