Rework USB OUT endpoint interrupts
Previously, when the USB interrupt was unmasked and a host issued an OUT
transfer to an endpoint that was not `read()` before the next OUT
transfer attempt to that endpoint, the USB ISR would continually fire
until `read()` was called on that endpoint.  If a higher priority ISR
didn't somehow clear the condition that prevented the `read()`, the CPU
would lock up.

This change moves the clearing of the endpoint's transfer complete and
transfer failed interrupt flags in to `poll()`, and enables the transfer
failed interrupt flag.  The transfer failed interrupt is important
because because when the host attempts an OUT transfer but the device
hasn't cleared the previous data, the device will NAK (assuming the
endpoint is not isochronous) and the host will usually retry a short
time later.  The NAK will set the transfer failed flag, which gives the
USB Class another opportunity to call `read()`.

The net effect is that a USB Class can safely skip a `read()` without
having to chose between discarding data, or locking up in the USB ISR.