Remove some unsafe and improve documentation for gpio::v2
The `RegisterInterface` trait in `gpio::v2` used an approach that
produced an unnecessary amount of `unsafe` code. Because each register
in a PAC struct is wrapped in an `UnsafeCell`, it is safe to mutate them
through shared references. The previous approach used a combination of
raw pointers and unique references. However, raw pointers are
unnecessary and unique references offer no benefits, yet they carry the
risk of UB if not implemented correctly.

Update the `RegisterInterface` trait to use shared references
throughout. Also improve the `pin` module documentation slightly.