Remove unused panic-abort and unnecessary panic-halt feature/dependency
Some boards have default panic-halt feature and dependency, but this doesn't make sense for users of the BSPs. This dependency is a dev-dependency for the examples. Also, consistency around examples using halt or semihosting panic option was varried on the boards that had this issue, so tried to clean that up as well.