clock/v2: Rename identity types for consistency
Several pairs of existing types in the `clock` module have the same
relationship as `PinId` and `DynPinId`. Rename these types to show the
symmetry.

Also rename all `marker::*` types to end with an `Id` suffix and lift
them out of the `marker` modules. Add documentation to explain that
these types represent the clock *identity*, not the clock itself. This
is simpler and easier to understand than the existing `marker` approach.
Rename the `*SourceMarker` traits to `*SourceId` as well.