clock/v2: Refactor `dpll` module
Refactor the `dpll` module to remove the `*Driven` types and use `*Id`
types for the second type parameter. In doing so, reimplement
`PclkSourceId` and `DynPclkSourceId` in terms of `GclkId` and a new
`DynGclkId`. Make the `Pclk` types effectively aliases for the `Gclk`
types, since a `Pclk` can only be sourced by a `Gclk`. Remove the
`dpll::SrcMode` trait and move elements of it to the `DpllSourceId`
trait. Refactor the `Dpll` methods to be compatible with the new trait
structure. Also update some documentation throughout.