clock/v2: Add `Enabled` type aliases
The documentation for methods defined on `Enabled<T>` for some clock
type `T` normally appear on the `Enabled` type. This is reasonable, but
for the `clock` module, we would prefer to separate the documentation
depending on the inner type `T`. The easiest way to accomplish this is
to add a series of type aliases. Rustdoc does a nice job of showing
methods defined on type aliases on their own, dedicated page. This makes
the documentation much more comprehensible and discoverable.

In the process, update `seq_macro` to v0.3, which makes it more
compatible with the 2021 Edition.