clock_v2: Refactor `ahb` and `apb` modules
The existing approach for the AHB and APB clocks is not sound, because
it could allow simultaneous modification of the same register from
different executation contexts.

Refactor both modules to use a central, singleton structure for register
access. Converting from a `Token` to a `Clk` and back requires `&mut`
access to the structure, which guarantees exclusive access.