Improve the architecture of `SpiFuture`
After trying to use `SpiFuture` in the real world, I encountered some
problems that these changes address. Specifically, I added the ability
for the `Spi` struct to be part of some user type. The `AsSpi` trait
provides an interface for the `SpiFuture`. This has the added benefit of
simplifying the implementation.