Fix text baseline broken in recent version bump
As I noted in #603, PR #591 bumped a bunch of crates across what are
effectively major version boundaries, because the crates are in the
leading-zero phase of semver -- so an 0.6 to 0.7 update is breaking.
This has broken _at least_ text rendering.

This commit fixes the text baseline calculations to use the method
expected by the embedded-graphics 0.7 crate, which has (quite
reasonably) switched to using baselines instead of bounding box top to
place text, but opted to leave the old implicit-baseline `new` operation
in place and change its behavior, which allowed all of this to break
without any compile errors.

NOTE THAT WITH THIS COMMIT NEARLY ALL OF THE WIO TERMINAL DEMOS ARE
STILL BADLY BROKEN. I'm new here and I'm not sure if these worked before
the version bump, or if they were also broken by the merge of this
untested PR, but at this point, only clock and usb_serial_display work.
Everything else displays blinking garbage on the display and hangs.