thumbv6m: fix USB clock recovery
USB clock recovery was disabled because it caused USB resets and instability.
This occurred because the DFFLMUL.MUL was set to the wrong value. The
multiplier must be 48000 to obtain a 48 MHz clock from the 1 kHz USB SOF
signal.

There appears to be no harm in unconditionally enabling USB clock recovery
when there is no external crystal, since it will just operate in open loop
mode (as before) when USB is not connected.