Clocking API V2 (for thumbv7em)  (#450)
Add a new API for the `clock` module on `thumbv7em` targets

Add a new `clock` module API that enforces correctness by construction.
It is impossible to create an invalid clock tree without using `unsafe`.
Specifically, use typed tokens and clocks to guarantee memory safety and
act as proof that a clock is correctly configured, and use type-level
numbers to count consumer clocks at compile-time and restrict a given
clock's API while it is in use.

This commit comes after two years of work, starting with #272, then
with #429 and culminating with PR #450.

Co-authored-by: Bradley Harden <bradleyharden@gmail.com>
Co-authored-by: Gabriel Górski <gabriel.gorski@volvocars.com>
Co-authored-by: Henrik Tjäder <henrik.tjader@volvocars.com>