Refactor Cargo features (#656)
Refactor the Cargo features to provide a better, more fine-grained
structure. Separate the features for selecting pins from the features
for selecting peripherals. Reorganize and better document the Cargo.toml
file.

The previous update scripts assumed a particular structure of the TOML
files. This approach is extremely brittle in the face of formatting or
comments. Instead, use the Python package `tomlkit` to parse, modify and
dump TOML files while still preserving all formatting and commenting.