fix wio_terminal: Upgrade heapless and nom
This resolves a dependabot alert for CVE-2020-36464 in heapless, as used
in the wio_terminal examples, also brings in support for const generics
so example code is a bit cleaner.

Current stable rustc warns about the older version of nom using features
which will be removed in future versions of Rust, so bump that too.