GCLK refactor
Reading the GCLK GENCTRL register isn't directly possible on the
thumbv6m chips, on these there is only one GENCTRL register for the
whole GCLK peripheral rather than one GENCTRL per generator.  On the
thumbv6m chips, modifying GENCTRL is done by atomically writing the
whole register, which has a field for the generator ID to be modified.

The GENCTRL register had been modified from the GclkToken<> impl,
per-field, but the state required to fill the other fields for the
atomic register write wasn't available in the GclkToken<> scope.  So,
these GENCTRL accesses were moved up in to the Gclk<> impl.

One more bit of state was added to the settings to handle the Output
Enable bit.