build: Update circuit-playground-express dependencies.
I updated to the latest version of the atsamd-hal. A number of other
deps needed to be updated to get there.

I also updated all the examples to compile. Examples for uart and usb
serial needed substantial updates so I just copied the version from the
feather_m0 crate and ported them.

I tested the following examples:
* blinky_basic
* neopixel_rainbow
* usb_serial

I wasn't able to test the uart example, but it does compile.

Closes: #951